Update below,
1. Add support to 8-digit checksum
2. Change time format to 24hr clock
3. Update App version to 1.3.1726.1
	Major Version: 1
	Minor Version: 3
	Revision: 1726 (means 2017 26th week)
	Build: 1 (means Monday, PS. 0 means Sunday)

+++ Source/2Checksum/ChecksumLib.cs
@@ -20,7 +20,7 @@ namespace ChecksumLib
         //
         // Properties
         //
-        private UInt16 _Checksum = 0;
+        private UInt32 _Checksum = 0;
 
         //
         // Structures
@@ -34,7 +34,7 @@ namespace ChecksumLib
         //
         // Procedure: CalcChecksum
         //
-        public void CalcChecksum(string Filename, out UInt16 Checksum)
+        public void CalcChecksum(string Filename, out UInt32 Checksum)
         {
             Byte[] Buffer;
             Byte[][] PartialBuffers;
@@ -108,7 +108,7 @@ namespace ChecksumLib
         private void PartialChecksumCalcThread(object Obj)
         {
             PartialChecksumCalcThreadParam Param = (PartialChecksumCalcThreadParam)Obj;
-            UInt16 PartialChecksum = 0;
+            UInt32 PartialChecksum = 0;
 
             for (int i = 0; i < Param.StreamSize; i++)
             {
